requirements.txt
pytest~=7.4.0
Flask~=2.3.2
Flask-WTF~=1.1.1
password-validator~=1.0
werkzeug~=2.3.7
WTForms~=3.0.1
python-dotenv~=1.0.0
SQLAlchemy~=1.4.41